+++ src/main/java/com/yo/day1/dto/payment/PaymentCreateRequest.java
@@ -9,7 +9,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDateTime;
 
 @Data
@@ -26,7 +26,7 @@ public class PaymentCreateRequest {
 
     @NotNull
     @Positive
-    private BigDecimal paidAmount;
+    private float paidAmount;
 
     @NotNull
     private PaymentMethod paymentMethod;
