+++ src/main/java/com/yo/day1/dto/payment/PaymentUpdateRequest.java
@@ -7,7 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDateTime;
 
 @Data
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 public class PaymentUpdateRequest {
 
     @Positive
-    private BigDecimal paidAmount;
+    private Float paidAmount;
 
     private PaymentMethod paymentMethod;
